+++ lib/src/isolate.dart
@@ -492,6 +492,13 @@ class VMStep {
   /// The isolate takes a single step, skipping over function calls.
   static const over = const VMStep._("Over");
 
+  /// The isolate continues until the execution returns from the an await, yield,
+  /// or yield* statement.
+  ///
+  /// Note that this value is only valid from VM service version 3.3 and when
+  /// [VMPauseEvent.atAsyncSuspension] is true.
+  static const overAsyncSuspension = const VMStep._("OverAsyncSuspension");
+
   /// The isolate continues until it exits the current function.
   static const out = const VMStep._("Out");